code-challenge/Answers/requirements.txt
mysql-connector-python
numpy
pandas
etherscan-python
jinja2